refactor: Extract static_files tests to separate file
- Moved 2,147 lines of tests from static_files.rs to static_files_test.rs
- Reduced main file from 2,420 to 273 lines (88.6% reduction)
- Preserved all test functionality and mock implementations
- Part of issue #118: improving codebase navigability
